Split api and normal files, codacy fixes (#1764)
* Split api and normal files, codacy fixes


* Codacy Fixes
