+++ tests/ha/multi_pod_failure/test_multi_data_pod_failure.py
@@ -22,7 +22,6 @@
 HA test suite for Multiple (K) Data Pods Failure
 """
 
-import copy
 import logging
 import os
 import secrets
@@ -45,8 +44,6 @@ from config.s3 import S3_CFG
 from libs.di.di_mgmt_ops import ManagementOPs
 from libs.ha.ha_common_libs_k8s import HAK8s
 from libs.prov.prov_k8s_cortx_deploy import ProvDeployK8sCortxLib
-from libs.s3.s3_blackbox_test_lib import JCloudClient
-from libs.s3.s3_multipart_test_lib import S3MultipartTestLib
 from libs.s3.s3_rest_cli_interface_lib import S3AccountOperations
 from libs.s3.s3_test_lib import S3TestLib
 
@@ -67,8 +64,6 @@ class TestMultiDataPodFailure:
         Setup operations for the test file.
         """
         LOGGER.info("STARTED: Setup Module operations.")
-        cls.csm_user = CMN_CFG["csm"]["csm_admin_user"]["username"]
-        cls.csm_passwd = CMN_CFG["csm"]["csm_admin_user"]["password"]
         cls.username = list()
         cls.password = list()
         cls.node_master_list = list()
@@ -80,7 +75,7 @@ class TestMultiDataPodFailure:
         cls.s3_clean = cls.test_prefix = None
         cls.s3acc_name = cls.s3acc_email = cls.bucket_name = cls.object_name = None
         cls.restore_pod = cls.deployment_backup = cls.deployment_name = cls.restore_method = None
-        cls.node_name = cls.deploy = cls.kvalue = cls.multipart_obj_path = None
+        cls.node_name = cls.deploy = cls.kvalue = None
         cls.pod_dict = dict()
         cls.mgnt_ops = ManagementOPs()
         cls.system_random = secrets.SystemRandom()
@@ -102,8 +97,6 @@ class TestMultiDataPodFailure:
                                                         password=cls.password[node]))
 
         cls.rest_obj = S3AccountOperations()
-        cls.s3_mp_test_obj = S3MultipartTestLib(endpoint_url=S3_CFG["s3_url"])
-        cls.test_file = "ha-mp_obj"
         cls.test_dir_path = os.path.join(TEST_DATA_FOLDER, "HATestMultipartUpload")
 
     def setup_method(self):
@@ -135,7 +128,6 @@ class TestMultiDataPodFailure:
         self.deployment_backup = None
         if not os.path.exists(self.test_dir_path):
             sysutils.make_dirs(self.test_dir_path)
-        self.multipart_obj_path = os.path.join(self.test_dir_path, self.test_file)
         LOGGER.info("Done: Setup operations.")
 
     def teardown_method(self):
@@ -224,7 +216,7 @@ class TestMultiDataPodFailure:
             down_method=const.RESTORE_DEPLOYMENT_K8S, kvalue=self.kvalue)
         # Assert if empty dictionary
         assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        for pod_name in resp[1].keys():
+        for pod_name in resp[1]:
             pod_data = list()
             pod_data.append(resp[1][pod_name]['hostname'])  # hostname
             pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
@@ -275,7 +267,7 @@ class TestMultiDataPodFailure:
                 master_node_obj=self.node_master_list[0], health_obj=self.hlth_master_list[0],
                 down_method=const.RESTORE_DEPLOYMENT_K8S)
             assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-            pod_name = list(resp[1].keys())[0]
+            pod_name = list(resp[1])[0]
             pod_data = list()
             pod_data.append(resp[1][pod_name]['hostname'])  # hostname
             pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
@@ -445,7 +437,7 @@ class TestMultiDataPodFailure:
         LOGGER.info("Starting three independent background threads for READs, WRITEs & DELETEs.")
         LOGGER.info("Step 3: Start Continuous DELETEs in background on random %s buckets",
                     del_bucket)
-        bucket_list = list(s3_data.keys())
+        bucket_list = list(s3_data)
         get_random_buck = self.system_random.sample(bucket_list, del_bucket)
         del_random_buck = get_random_buck.copy()
         args = {'test_prefix': test_prefix_del, 'test_dir_path': self.test_dir_path,
@@ -497,7 +489,7 @@ class TestMultiDataPodFailure:
             kvalue=self.kvalue, event=event, event_set_clr=event_set_clr)
         # Assert if empty dictionary
         assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        for pod_name in resp[1].keys():
+        for pod_name in resp[1]:
             pod_data = list()
             pod_data.append(resp[1][pod_name]['hostname'])  # hostname
             pod_data.append(resp[1][pod_name]['deployment_name'])  # deployment_name
@@ -588,119 +580,6 @@ class TestMultiDataPodFailure:
 
     @pytest.mark.ha
     @pytest.mark.lc
-    @pytest.mark.tags("TEST-35793")
-    def test_chunk_upload_during_data_kpods_down(self):
-        """
-        Test chunk upload during k data pods going down using delete deployment (using jclient)
-        """
-        LOGGER.info("STARTED: Test chunk upload during K data pods going down by delete "
-                    "deployment (using jclient)")
-        file_size = HA_CFG["5gb_mpu_data"]["file_size"]
-        download_file = "test_chunk_upload" + "_download"
-        download_path = os.path.join(self.test_dir_path, download_file)
-        chunk_obj_path = os.path.join(self.test_dir_path, self.object_name)
-        output = Queue()
-        event = threading.Event()
-
-        LOGGER.info("Step 1: Perform setup steps for jclient")
-        jc_obj = JCloudClient()
-        resp = self.ha_obj.setup_jclient(jc_obj)
-        assert_utils.assert_true(resp, "Failed in setting up jclient")
-        LOGGER.info("Step 1: Successfully setup jcloud/jclient on runner")
-
-        LOGGER.info("Step 2: Create IAM user with name %s, bucket %s and start chunk upload in "
-                    "background", self.s3acc_name, self.bucket_name)
-        resp = self.rest_obj.create_s3_account(acc_name=self.s3acc_name,
-                                               email_id=self.s3acc_email,
-                                               passwd=S3_CFG["CliConfig"]["s3_account"]["password"])
-        assert_utils.assert_true(resp[0], resp[1])
-        access_key = resp[1]["access_key"]
-        secret_key = resp[1]["secret_key"]
-        self.s3_clean = {'s3_acc': {'accesskey': access_key, 'secretkey': secret_key,
-                                    'user_name': self.s3acc_name}}
-        args = {'s3_data': self.s3_clean, 'bucket_name': self.bucket_name,
-                'file_size': file_size, 'chunk_obj_path': chunk_obj_path, 'output': output}
-
-        thread = threading.Thread(target=self.ha_obj.create_bucket_chunk_upload, kwargs=args)
-        thread.daemon = True  # Daemonize thread
-        thread.start()
-
-        LOGGER.info("Step 2: Successfully started chuck upload in background. Sleeping for %s "
-                    "sec", HA_CFG["common_params"]["30sec_delay"])
-        time.sleep(HA_CFG["common_params"]["30sec_delay"])
-
-        LOGGER.info("Step 3: Shutdown random %s (K) data pods by deleting deployment and "
-                    "verify cluster & remaining pods status", self.kvalue)
-        resp = self.ha_obj.delete_kpod_with_shutdown_methods(
-            master_node_obj=self.node_master_list[0], health_obj=self.hlth_master_list[0],
-            kvalue=self.kvalue, event=event, down_method=const.RESTORE_DEPLOYMENT_K8S)
-        # Assert if empty dictionary
-        assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        for pod_name in resp[1].keys():
-            pod_data = list()
-            pod_data.append(resp[1][pod_name]['hostname'])  # hostname
-            pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
-            pod_data.append(resp[1][pod_name]['deployment_name'])  # deployment_name
-            self.pod_name_list.append(pod_name)
-            self.pod_dict[pod_name] = pod_data
-        self.restore_pod = self.deploy = True
-        self.restore_method = const.RESTORE_DEPLOYMENT_K8S
-        assert_utils.assert_true(resp[0], "Cluster/Services status is not as expected")
-        LOGGER.info("Step 3: Successfully shutdown %s (K) data pod %s. Verified cluster and "
-                    "services states are as expected & remaining pods status is online.",
-                    self.kvalue, self.pod_name_list)
-
-        LOGGER.info("Step 4: Verifying response of background process")
-        thread.join()
-        while True:
-            resp = output.get(timeout=HA_CFG["common_params"]["60sec_delay"])
-            if isinstance(resp, bool):
-                break
-        if resp is None:
-            assert_utils.assert_true(False, "Background process of chunk upload failed")
-        LOGGER.info("Step 4: Successfully verified response of background process")
-
-        if not resp:
-            LOGGER.info("Step 5: Chunk upload failed in between, trying chunk upload again")
-            self.ha_obj.create_bucket_chunk_upload(s3_data=self.s3_clean,
-                                                   bucket_name=self.bucket_name,
-                                                   file_size=file_size,
-                                                   chunk_obj_path=chunk_obj_path,
-                                                   output=output,
-                                                   bkt_op=False)
-
-            while True:
-                resp = output.get(timeout=HA_CFG["common_params"]["60sec_delay"])
-                if isinstance(resp, bool):
-                    break
-
-            if not resp or resp is None:
-                assert_utils.assert_true(False, "Retried chunk upload failed")
-            LOGGER.info("Step 5: Retried chunk upload completed successfully")
-
-        LOGGER.info("Calculating checksum of file %s", chunk_obj_path)
-        upload_checksum = self.ha_obj.cal_compare_checksum(file_list=[chunk_obj_path],
-                                                           compare=False)[0]
-
-        LOGGER.info("Step 6: Download object and verify checksum")
-        resp = self.ha_obj.object_download_jclient(s3_data=self.s3_clean,
-                                                   bucket_name=self.bucket_name,
-                                                   object_name=self.object_name,
-                                                   obj_download_path=download_path)
-        LOGGER.info("Download object response: %s", resp)
-        assert_utils.assert_true(resp[0], resp[1])
-        download_checksum = self.ha_obj.cal_compare_checksum(file_list=[download_path],
-                                                             compare=False)[0]
-        assert_utils.assert_equal(upload_checksum, download_checksum,
-                                  f"Failed to match checksum: {upload_checksum},"
-                                  f" {download_checksum}")
-        LOGGER.info("Step 6: Successfully downloaded object and verified checksum")
-
-        LOGGER.info("ENDED: Test chunk upload during k data pods going down by delete "
-                    "deployment (unsafe)")
-
-    @pytest.mark.ha
-    @pytest.mark.lc
     @pytest.mark.tags("TEST-35777")
     def test_reads_writes_during_kpods_down(self):
         """
@@ -760,7 +639,7 @@ class TestMultiDataPodFailure:
             event_set_clr=event_set_clr)
         # Assert if empty dictionary
         assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        for pod_name in resp[1].keys():
+        for pod_name in resp[1]:
             pod_data = list()
             pod_data.append(resp[1][pod_name]['hostname'])  # hostname
             pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
@@ -886,7 +765,7 @@ class TestMultiDataPodFailure:
             down_method=const.RESTORE_DEPLOYMENT_K8S, kvalue=self.kvalue)
         # Assert if empty dictionary
         assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        for pod_name in resp[1].keys():
+        for pod_name in resp[1]:
             pod_data = list()
             pod_data.append(resp[1][pod_name]['hostname'])  # hostname
             pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
@@ -980,7 +859,7 @@ class TestMultiDataPodFailure:
             event_set_clr=event_set_clr)
         # Assert if empty dictionary
         assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        for pod_name in resp[1].keys():
+        for pod_name in resp[1]:
             pod_data = list()
             pod_data.append(resp[1][pod_name]['hostname'])  # hostname
             pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
@@ -1057,7 +936,7 @@ class TestMultiDataPodFailure:
             event_set_clr=event_set_clr)
         # Assert if empty dictionary
         assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        for pod_name in resp[1].keys():
+        for pod_name in resp[1]:
             pod_data = list()
             pod_data.append(resp[1][pod_name]['hostname'])  # hostname
             pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
@@ -1168,7 +1047,7 @@ class TestMultiDataPodFailure:
             kvalue=self.kvalue, event=event, down_method=const.RESTORE_DEPLOYMENT_K8S)
         # Assert if empty dictionary
         assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        for pod_name in resp[1].keys():
+        for pod_name in resp[1]:
             pod_data = list()
             pod_data.append(resp[1][pod_name]['hostname'])  # hostname
             pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
@@ -1259,7 +1138,7 @@ class TestMultiDataPodFailure:
         LOGGER.info("ENDED: Test to verify DELETEs while pods are failing till K "
                     "data pods are failed.")
 
-    # pylint: disable=C0321
+    # pylint: disable=multiple-statements
     # pylint: disable=too-many-locals
     @pytest.mark.ha
     @pytest.mark.lc
@@ -1313,7 +1192,7 @@ class TestMultiDataPodFailure:
         LOGGER.info("Starting three independent background threads for READs, WRITEs & DELETEs.")
         LOGGER.info("Step 3: Start Continuous DELETEs in background on random %s buckets",
                     del_bucket)
-        bucket_list = list(s3_data.keys())
+        bucket_list = list(s3_data)
         get_random_buck = self.system_random.sample(bucket_list, del_bucket)
         args = {'test_prefix': test_prefix_del, 'test_dir_path': self.test_dir_path,
                 'skipput': True, 'skipget': True, 'bkt_list': get_random_buck, 'output': del_output}
@@ -1370,7 +1249,7 @@ class TestMultiDataPodFailure:
             self.restore_method = const.RESTORE_DEPLOYMENT_K8S
             self.pod_dict[pod_name] = pod_data
             LOGGER.info("Deleted %s data pod %s by deleting deployment (unsafe)", count, pod_name)
-        LOGGER.info("Step 6: Sucessfully shutdown %s (K) data pods one by one while continuous "
+        LOGGER.info("Step 6: Successfully shutdown %s (K) data pods one by one while continuous "
                     "IOs in background", self.kvalue)
         event.clear()
 
@@ -1497,7 +1376,7 @@ class TestMultiDataPodFailure:
                 master_node_obj=self.node_master_list[0], health_obj=self.hlth_master_list[0],
                 down_method=const.RESTORE_DEPLOYMENT_K8S)
             assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-            pod_name = list(resp[1].keys())[0]
+            pod_name = list(resp[1])[0]
             pod_data = list()
             pod_data.append(resp[1][pod_name]['hostname'])  # hostname
             pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
@@ -1565,7 +1444,7 @@ class TestMultiDataPodFailure:
             down_method=const.RESTORE_DEPLOYMENT_K8S, kvalue=self.kvalue)
         # Assert if empty dictionary
         assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        for pod_name in resp[1].keys():
+        for pod_name in resp[1]:
             pod_data = list()
             pod_data.append(resp[1][pod_name]['hostname'])  # hostname
             pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
@@ -1657,7 +1536,7 @@ class TestMultiDataPodFailure:
                 master_node_obj=self.node_master_list[0], health_obj=self.hlth_master_list[0],
                 down_method=const.RESTORE_DEPLOYMENT_K8S)
             assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-            pod_name = list(resp[1].keys())[0]
+            pod_name = list(resp[1])[0]
             pod_data = list()
             pod_data.append(resp[1][pod_name]['hostname'])  # hostname
             pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
@@ -1720,703 +1599,3 @@ class TestMultiDataPodFailure:
 
         LOGGER.info("Completed: Test to verify degraded DELETEs after each pod failure till K "
                     "data pods fail.")
-
-    @pytest.mark.ha
-    @pytest.mark.lc
-    @pytest.mark.tags("TEST-35785")
-    def test_copy_object_kpods_fail(self):
-        """
-        Test to Verify copy object when all K data pods are failed.
-        """
-        LOGGER.info("STARTED: Test to Verify copy object when all K data pods are failed.")
-
-        bkt_cnt = HA_CFG["copy_obj_data"]["bkt_cnt"]
-        bkt_obj_dict = dict()
-        for cnt in range(bkt_cnt):
-            t_t = int(perf_counter_ns())
-            bkt_obj_dict[f"ha-bkt-{cnt}-{t_t}"] = f"ha-obj-{cnt}-{t_t}"
-        event = threading.Event()
-        LOGGER.info("Creating IAM user with name %s", self.s3acc_name)
-        resp = self.rest_obj.create_s3_account(acc_name=self.s3acc_name,
-                                               email_id=self.s3acc_email,
-                                               passwd=S3_CFG["CliConfig"]["s3_account"]["password"])
-        assert_utils.assert_true(resp[0], resp[1])
-        access_key = resp[1]["access_key"]
-        secret_key = resp[1]["secret_key"]
-        s3_test_obj = S3TestLib(access_key=access_key, secret_key=secret_key,
-                                endpoint_url=S3_CFG["s3_url"])
-        LOGGER.info("Successfully created IAM user with name %s", self.s3acc_name)
-        self.s3_clean = {'s3_acc': {'accesskey': access_key, 'secretkey': secret_key,
-                                    'user_name': self.s3acc_name}}
-        LOGGER.info("Step 1: Create and list buckets and perform upload and copy "
-                    "object from %s bucket to other buckets ", self.bucket_name)
-        resp = self.ha_obj.create_bucket_copy_obj(event, s3_test_obj=s3_test_obj,
-                                                  bucket_name=self.bucket_name,
-                                                  object_name=self.object_name,
-                                                  bkt_obj_dict=bkt_obj_dict,
-                                                  file_path=self.multipart_obj_path)
-        assert_utils.assert_true(resp[0], f"Failed buckets are: {resp[1]}")
-        put_etag = resp[1]
-        LOGGER.info("Step 1: successfully create and list buckets and perform upload and copy"
-                    "object from %s bucket to other buckets", self.bucket_name)
-
-        LOGGER.info("Step 2: Shutdown random %s (K) data pods by deleting deployment and "
-                    "verify cluster & remaining pods status", self.kvalue)
-        resp = self.ha_obj.delete_kpod_with_shutdown_methods(
-            master_node_obj=self.node_master_list[0], health_obj=self.hlth_master_list[0],
-            down_method=const.RESTORE_DEPLOYMENT_K8S, kvalue=self.kvalue)
-        # Assert if empty dictionary
-        assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        for pod_name in resp[1].keys():
-            pod_data = list()
-            pod_data.append(resp[1][pod_name]['hostname'])  # hostname
-            pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
-            pod_data.append(resp[1][pod_name]['deployment_name'])  # deployment_name
-            self.pod_name_list.append(pod_name)
-            self.pod_dict[pod_name] = pod_data
-        self.restore_pod = self.deploy = True
-        self.restore_method = const.RESTORE_DEPLOYMENT_K8S
-        assert_utils.assert_true(resp[0], "Cluster/Services status is not as expected")
-        LOGGER.info("Step 2: Successfully shutdown %s (K) data pod %s. Verified cluster and "
-                    "services states are as expected & remaining pods status is online.",
-                    self.kvalue, self.pod_name_list)
-
-        LOGGER.info("Step 3: Download the uploaded objects & verify etags")
-        for key, val in bkt_obj_dict.items():
-            resp = s3_test_obj.get_object(bucket=key, key=val)
-            LOGGER.info("Get object response: %s", resp)
-            get_etag = resp[1]["ETag"]
-            assert_utils.assert_equal(put_etag, get_etag, "Failed in Etag verification of "
-                                                          f"object {key} of bucket {val}. "
-                                                          "Put and Get Etag mismatch")
-        LOGGER.info("Step 3: Successfully download the uploaded objects & verify etags")
-
-        t_t = int(perf_counter_ns())
-        bucket_name = self.bucket_name
-        bkt_op = False
-        if CMN_CFG["dtm0_disabled"]:
-            LOGGER.info("Create and list buckets")
-            bkt_obj_dict.clear()
-            for cnt in range(bkt_cnt):
-                bkt_obj_dict[f"ha-bkt{cnt}-{t_t}"] = f"ha-obj{cnt}-{t_t}"
-            bucket_name = f"ha-mp-bkt-{t_t}-2"
-            bkt_op = True
-        else:
-            for idx, bkt in enumerate(bkt_obj_dict):
-                bkt_obj_dict[bkt] = f"ha-obj{idx}-{t_t}"
-
-        LOGGER.info("Step 4: Perform copy object from %s bucket to other buckets verify copy object"
-                    " etags", bucket_name)
-        resp = self.ha_obj.create_bucket_copy_obj(event, s3_test_obj=s3_test_obj,
-                                                  bucket_name=bucket_name,
-                                                  object_name=self.object_name,
-                                                  bkt_obj_dict=bkt_obj_dict,
-                                                  put_etag=put_etag,
-                                                  bkt_op=bkt_op)
-        assert_utils.assert_true(resp[0], f"Failed buckets are: {resp[1]}")
-        LOGGER.info("Step 4: Performed copy object from %s bucket to other buckets verified copy "
-                    "object etags", bucket_name)
-
-        LOGGER.info("Step 5: Download the copied objects & verify etags.")
-        for bkt, obj in bkt_obj_dict.items():
-            resp = s3_test_obj.get_object(bucket=bkt, key=obj)
-            LOGGER.info("Get object response: %s", resp)
-            get_etag = resp[1]["ETag"]
-            assert_utils.assert_equal(put_etag, get_etag, "Failed in verification of Put & Get "
-                                                          f"Etag for object {obj} of bucket "
-                                                          f"{bkt}.")
-        LOGGER.info("Step 5: Downloaded copied objects & verify etags.")
-
-        LOGGER.info("COMPLETED: Test to Verify copy object when all K data pods are failed.")
-
-    @pytest.mark.ha
-    @pytest.mark.lc
-    @pytest.mark.tags("TEST-35782")
-    def test_degraded_mpu_after_kpods_fail(self):
-        """
-        This test tests multipart upload after all K data pods are failed
-        """
-        LOGGER.info("STARTED: Test to verify multipart upload after all K data pods are failed.")
-        file_size = HA_CFG["5gb_mpu_data"]["file_size"]
-        total_parts = HA_CFG["5gb_mpu_data"]["total_parts"]
-        download_file = self.test_file + "_download"
-        download_path = os.path.join(self.test_dir_path, download_file)
-
-        LOGGER.info("Step 1: Create bucket and perform multipart upload of size %sMB.",
-                    file_size)
-        LOGGER.info("Creating IAM user with name %s", self.s3acc_name)
-        resp = self.rest_obj.create_s3_account(acc_name=self.s3acc_name,
-                                               email_id=self.s3acc_email,
-                                               passwd=S3_CFG["CliConfig"]["s3_account"]["password"])
-        assert_utils.assert_true(resp[0], resp[1])
-        access_key = resp[1]["access_key"]
-        secret_key = resp[1]["secret_key"]
-        s3_test_obj = S3TestLib(access_key=access_key, secret_key=secret_key,
-                                endpoint_url=S3_CFG["s3_url"])
-        LOGGER.info("Successfully created IAM user with name %s", self.s3acc_name)
-        self.s3_clean = {'s3_acc': {'accesskey': access_key, 'secretkey': secret_key,
-                                    'user_name': self.s3acc_name}}
-        resp = self.ha_obj.create_bucket_to_complete_mpu(s3_data=self.s3_clean,
-                                                         bucket_name=self.bucket_name,
-                                                         object_name=self.object_name,
-                                                         file_size=file_size,
-                                                         total_parts=total_parts,
-                                                         multipart_obj_path=self.multipart_obj_path)
-        assert_utils.assert_true(resp[0], resp)
-        result = s3_test_obj.object_info(self.bucket_name, self.object_name)
-        obj_size = result[1]["ContentLength"]
-        LOGGER.debug("Uploaded object of size %s for %s", obj_size, self.bucket_name)
-        assert_utils.assert_equal(obj_size, file_size * const.Sizes.MB)
-        upload_checksum = str(resp[2])
-        LOGGER.info("Step 1: Successfully performed multipart upload for size size %sMB.",
-                    file_size)
-
-        LOGGER.info("Step 2: Shutdown random %s (K) data pods by deleting deployment and "
-                    "verify cluster & remaining pods status", self.kvalue)
-        resp = self.ha_obj.delete_kpod_with_shutdown_methods(
-            master_node_obj=self.node_master_list[0], health_obj=self.hlth_master_list[0],
-            down_method=const.RESTORE_DEPLOYMENT_K8S, kvalue=self.kvalue)
-        # Assert if empty dictionary
-        assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        for pod_name in resp[1].keys():
-            pod_data = list()
-            pod_data.append(resp[1][pod_name]['hostname'])  # hostname
-            pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
-            pod_data.append(resp[1][pod_name]['deployment_name'])  # deployment_name
-            self.pod_name_list.append(pod_name)
-            self.pod_dict[pod_name] = pod_data
-        self.restore_pod = self.deploy = True
-        self.restore_method = const.RESTORE_DEPLOYMENT_K8S
-        assert_utils.assert_true(resp[0], "Cluster/Services status is not as expected")
-        LOGGER.info("Step 2: Successfully shutdown %s (K) data pod %s. Verified cluster and "
-                    "services states are as expected & remaining pods status is online.",
-                    self.kvalue, self.pod_name_list)
-
-        LOGGER.info("Step 3: Download the uploaded object and verify checksum")
-        resp = s3_test_obj.object_download(self.bucket_name, self.object_name, download_path)
-        assert_utils.assert_true(resp[0], resp[1])
-        download_checksum = self.ha_obj.cal_compare_checksum(file_list=[download_path],
-                                                             compare=False)[0]
-        assert_utils.assert_equal(upload_checksum, download_checksum,
-                                  f"Failed to match checksum: {upload_checksum},"
-                                  f" {download_checksum}")
-        LOGGER.info("Matched checksum: %s, %s", upload_checksum, download_checksum)
-        LOGGER.info("Step 3: Successfully downloaded the object and verified the checksum")
-
-        LOGGER.info("Removing files %s and %s", self.multipart_obj_path, download_path)
-        sysutils.remove_file(self.multipart_obj_path)
-        sysutils.remove_file(download_path)
-
-        LOGGER.info("Step 4: Do multipart upload and download the object & verify checksum.")
-        if CMN_CFG["dtm0_disabled"]:
-            LOGGER.info("Creating new bucket for new upload")
-            bucket_name = f"mp-bkt-{perf_counter_ns()}"
-        else:
-            bucket_name = self.bucket_name
-        object_name = f"mp-obj-{int(perf_counter_ns())}"
-        resp = self.ha_obj.create_bucket_to_complete_mpu(s3_data=self.s3_clean,
-                                                         bucket_name=bucket_name,
-                                                         object_name=object_name,
-                                                         file_size=file_size,
-                                                         total_parts=total_parts,
-                                                         multipart_obj_path=self.multipart_obj_path)
-        assert_utils.assert_true(resp[0], resp)
-        upload_checksum1 = resp[2]
-        result = s3_test_obj.object_info(bucket_name, object_name)
-        obj_size = result[1]["ContentLength"]
-        LOGGER.debug("Uploaded object info for %s is %s", bucket_name, result)
-        assert_utils.assert_equal(obj_size, file_size * const.Sizes.MB)
-
-        resp = s3_test_obj.object_download(bucket_name, object_name, download_path)
-        LOGGER.info("Download object response: %s", resp)
-        assert_utils.assert_true(resp[0], resp[1])
-        download_checksum1 = self.ha_obj.cal_compare_checksum(file_list=[download_path],
-                                                              compare=False)[0]
-        assert_utils.assert_equal(upload_checksum1, download_checksum1,
-                                  f"Failed to match checksum: {upload_checksum1},"
-                                  f" {download_checksum1}")
-        LOGGER.info("Matched checksum: %s, %s", upload_checksum1, download_checksum1)
-        LOGGER.info("Step 4: Did multipart upload & downloaded the object & verified the checksum.")
-
-        LOGGER.info("COMPLETED: Test to verify multipart upload after all K data pods are failed.")
-
-    @pytest.mark.ha
-    @pytest.mark.lc
-    @pytest.mark.tags("TEST-35786")
-    def test_copy_object_during_kpods_down(self):
-        """
-        Test to verify copy object during data pods failure till K pods.
-        """
-        LOGGER.info("STARTED: Verify copy object during data pods failure till K pods.")
-
-        bkt_obj_dict = dict()
-        output = Queue()
-        bkt_obj_dict[f"ha-bkt-{perf_counter_ns()}"] = f"ha-obj-{perf_counter_ns()}"
-        event = threading.Event()
-
-        LOGGER.info("Creating IAM user with name %s", self.s3acc_name)
-        resp = self.rest_obj.create_s3_account(acc_name=self.s3acc_name,
-                                               email_id=self.s3acc_email,
-                                               passwd=S3_CFG["CliConfig"]["s3_account"]["password"])
-        assert_utils.assert_true(resp[0], resp[1])
-        access_key = resp[1]["access_key"]
-        secret_key = resp[1]["secret_key"]
-        s3_test_obj = S3TestLib(access_key=access_key, secret_key=secret_key,
-                                endpoint_url=S3_CFG["s3_url"])
-        LOGGER.info("Successfully created IAM user with name %s", self.s3acc_name)
-        self.s3_clean = {'s3_acc': {'accesskey': access_key, 'secretkey': secret_key,
-                                    'user_name': self.s3acc_name}}
-
-        LOGGER.info("Step 1: Create bucket, upload an object and copy to the bucket")
-        # This is done just to get put_etag for further ops.
-        resp = self.ha_obj.create_bucket_copy_obj(event, s3_test_obj=s3_test_obj,
-                                                  bucket_name=self.bucket_name,
-                                                  object_name=self.object_name,
-                                                  bkt_obj_dict=bkt_obj_dict,
-                                                  file_path=self.multipart_obj_path)
-        assert_utils.assert_true(resp[0], resp[1])
-        put_etag = resp[1]
-        LOGGER.info("Step 1: Successfully created bucket, uploaded and copied an object "
-                    "to the bucket")
-        bkt_obj_dict.clear()
-
-        bkt_cnt = HA_CFG["copy_obj_data"]["bkt_multi"]
-        for cnt in range(bkt_cnt):
-            rd_time = perf_counter_ns()
-            s3_test_obj.create_bucket(f"ha-bkt{cnt}-{rd_time}")
-            bkt_obj_dict[f"ha-bkt{cnt}-{rd_time}"] = f"ha-obj{cnt}-{rd_time}"
-        LOGGER.info("Step 2: Create multiple buckets and copy object from %s to other buckets in "
-                    "background", self.bucket_name)
-        args = {'s3_test_obj': s3_test_obj, 'bucket_name': self.bucket_name,
-                'object_name': self.object_name, 'bkt_obj_dict': bkt_obj_dict, 'output': output,
-                'file_path': self.multipart_obj_path, 'background': True, 'bkt_op': False,
-                'put_etag': put_etag}
-        thread = threading.Thread(target=self.ha_obj.create_bucket_copy_obj, args=(event,),
-                                  kwargs=args)
-        thread.daemon = True  # Daemonize thread
-        thread.start()
-        LOGGER.info("Step 2: Successfully started background process for copy object")
-        # While loop to sync this operation with background thread to achieve expected scenario
-        LOGGER.info("Waiting for creation of %s buckets", bkt_cnt)
-        bkt_list = list()
-        timeout = time.time() + 60 * 3
-        while len(bkt_list) < bkt_cnt:
-            time.sleep(HA_CFG["common_params"]["20sec_delay"])
-            bkt_list = s3_test_obj.bucket_list()[1]
-            if timeout < time.time():
-                LOGGER.error("Bucket creation is taking longer than 3 mins")
-                assert_utils.assert_true(False, "Please check background process logs")
-        time.sleep(HA_CFG["common_params"]["20sec_delay"])
-
-        LOGGER.info("Step 3: Shutdown random %s (K) data pods by deleting deployment and "
-                    "verify cluster & remaining pods status", self.kvalue)
-        resp = self.ha_obj.delete_kpod_with_shutdown_methods(
-            master_node_obj=self.node_master_list[0], health_obj=self.hlth_master_list[0],
-            down_method=const.RESTORE_DEPLOYMENT_K8S, kvalue=self.kvalue, event=event)
-        # Assert if empty dictionary
-        assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        for pod_name in resp[1].keys():
-            pod_data = list()
-            pod_data.append(resp[1][pod_name]['hostname'])  # hostname
-            pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
-            pod_data.append(resp[1][pod_name]['deployment_name'])  # deployment_name
-            self.pod_name_list.append(pod_name)
-            self.pod_dict[pod_name] = pod_data
-        self.restore_pod = self.deploy = True
-        self.restore_method = const.RESTORE_DEPLOYMENT_K8S
-        assert_utils.assert_true(resp[0], "Cluster/Services status is not as expected")
-        LOGGER.info("Step 3: Successfully shutdown %s (K) data pod %s. Verified cluster and "
-                    "services states are as expected & remaining pods status is online.",
-                    self.kvalue, self.pod_name_list)
-
-        LOGGER.info("Step 4: Checking responses from background copy process")
-        thread.join()
-        responses = tuple()
-        while len(responses) < 3:
-            responses = output.get(timeout=HA_CFG["common_params"]["60sec_delay"])
-
-        if not responses:
-            assert_utils.assert_true(False, "Background process failed to do copy object")
-
-        put_etag = responses[0]
-        exp_fail_bkt_obj_dict = responses[1]
-        failed_bkts = responses[2]
-        LOGGER.debug("Responses received from background process:\nput_etag: "
-                     "%s\nexp_fail_bkt_obj_dict: %s\nfailed_bkts: %s", put_etag,
-                     exp_fail_bkt_obj_dict, failed_bkts)
-        if len(exp_fail_bkt_obj_dict) == 0 and len(failed_bkts) == 0:
-            LOGGER.info("Copy object operation for all the buckets completed successfully. ")
-        elif failed_bkts:
-            assert_utils.assert_true(False, "Failed to do copy object when cluster was in degraded "
-                                            f"state. Failed buckets: {failed_bkts}")
-        elif exp_fail_bkt_obj_dict:
-            LOGGER.info("Step 4.1: Retrying copy object to buckets %s",
-                        list(exp_fail_bkt_obj_dict.keys()))
-            resp = self.ha_obj.create_bucket_copy_obj(event, s3_test_obj=s3_test_obj,
-                                                      bucket_name=self.bucket_name,
-                                                      object_name=self.object_name,
-                                                      bkt_obj_dict=exp_fail_bkt_obj_dict,
-                                                      bkt_op=False, put_etag=put_etag)
-            assert_utils.assert_true(resp[0], f"Failed buckets are: {resp[1]}")
-            put_etag = resp[1]
-        LOGGER.info("Step 4: Successfully checked and completed background copy process")
-
-        LOGGER.info("Step 5: Download the uploaded objects & verify etags")
-        for key, val in bkt_obj_dict.items():
-            resp = s3_test_obj.get_object(bucket=key, key=val)
-            LOGGER.info("Get object response: %s", resp)
-            get_etag = resp[1]["ETag"]
-            assert_utils.assert_equal(put_etag, get_etag, "Failed in Etag verification of "
-                                                          f"object {key} of bucket {val}. "
-                                                          "Put and Get Etag mismatch")
-        LOGGER.info("Step 5: Successfully download the uploaded objects & verify etags")
-
-        t_t = int(perf_counter_ns())
-        bucket_name = self.bucket_name
-        bkt_op = False
-        if CMN_CFG["dtm0_disabled"]:
-            LOGGER.info("Create and list buckets")
-            bkt_obj_dict.clear()
-            for cnt in range(bkt_cnt):
-                bkt_obj_dict[f"ha-bkt{cnt}-{t_t}"] = f"ha-obj{cnt}-{t_t}"
-            bucket_name = f"ha-mp-bkt-{t_t}-2"
-            bkt_op = True
-        else:
-            for idx, bkt in enumerate(bkt_obj_dict):
-                bkt_obj_dict[bkt] = f"ha-obj{idx}-{t_t}"
-
-        LOGGER.info("Step 6: Perform copy object from %s bucket to other buckets verify copy object"
-                    " etags", bucket_name)
-        resp = self.ha_obj.create_bucket_copy_obj(event, s3_test_obj=s3_test_obj,
-                                                  bucket_name=bucket_name,
-                                                  object_name=self.object_name,
-                                                  bkt_obj_dict=bkt_obj_dict,
-                                                  put_etag=put_etag,
-                                                  bkt_op=bkt_op)
-        assert_utils.assert_true(resp[0], f"Failed buckets are: {resp[1]}")
-        LOGGER.info("Step 6: Performed copy object from %s bucket to other buckets verified copy "
-                    "object etags", bucket_name)
-
-        LOGGER.info("Step 7: Download the copied objects & verify etags.")
-        for bkt, obj in bkt_obj_dict.items():
-            resp = s3_test_obj.get_object(bucket=bkt, key=obj)
-            LOGGER.info("Get object response: %s", resp)
-            get_etag = resp[1]["ETag"]
-            assert_utils.assert_equal(put_etag, get_etag, "Failed in verification of Put & Get "
-                                                          f"Etag for object {obj} of bucket "
-                                                          f"{bkt}.")
-        LOGGER.info("Step 7: Downloaded copied objects & verify etags.")
-
-        LOGGER.info("COMPLETED: Verify copy object during data pods failure till K pods.")
-
-    @pytest.mark.ha
-    @pytest.mark.lc
-    @pytest.mark.tags("TEST-35783")
-    def test_partial_mpu_after_kpods_fail(self):
-        """
-        This test tests partial multipart upload after each data pod is failed till K
-        pods and complete upload after all K data pods are failed
-        """
-        LOGGER.info("STARTED: Test to verify partial multipart upload after each data pod is "
-                    "failed till K data pods and complete upload after all K data pods are failed")
-        file_size = HA_CFG["5gb_mpu_data"]["file_size"]
-        total_parts = self.kvalue * 5 + HA_CFG["5gb_mpu_data"]["total_parts"]
-        parts = list(range(1, total_parts + 1))
-        download_file = self.test_file + "_download"
-        download_path = os.path.join(self.test_dir_path, download_file)
-        if os.path.exists(self.multipart_obj_path):
-            os.remove(self.multipart_obj_path)
-        sysutils.create_file(self.multipart_obj_path, file_size)
-        LOGGER.info("Calculating checksum of file %s", self.multipart_obj_path)
-        upload_checksum = self.ha_obj.cal_compare_checksum(file_list=[self.multipart_obj_path],
-                                                           compare=False)[0]
-
-        LOGGER.info("Creating IAM user with name %s", self.s3acc_name)
-        resp = self.rest_obj.create_s3_account(acc_name=self.s3acc_name,
-                                               email_id=self.s3acc_email,
-                                               passwd=S3_CFG["CliConfig"]["s3_account"]["password"])
-        assert_utils.assert_true(resp[0], resp[1])
-        access_key = resp[1]["access_key"]
-        secret_key = resp[1]["secret_key"]
-        s3_test_obj = S3TestLib(access_key=access_key, secret_key=secret_key,
-                                endpoint_url=S3_CFG["s3_url"])
-        s3_mp_test_obj = S3MultipartTestLib(access_key=access_key, secret_key=secret_key,
-                                            endpoint_url=S3_CFG["s3_url"])
-        LOGGER.info("Successfully created IAM user with name %s", self.s3acc_name)
-        self.s3_clean = {'s3_acc': {'accesskey': access_key, 'secretkey': secret_key,
-                                    'user_name': self.s3acc_name}}
-        uploading_parts = self.system_random.sample(parts, 10)
-        LOGGER.info("Step 1: Perform partial multipart upload for %s parts out of total %s",
-                    uploading_parts, len(parts))
-        resp = self.ha_obj.partial_multipart_upload(s3_data=self.s3_clean,
-                                                    bucket_name=self.bucket_name,
-                                                    object_name=self.object_name,
-                                                    part_numbers=uploading_parts,
-                                                    multipart_obj_size=file_size,
-                                                    total_parts=total_parts,
-                                                    multipart_obj_path=self.multipart_obj_path)
-        mpu_id = resp[1]
-        object_path = resp[2]
-        parts_etags = copy.deepcopy(resp[3])
-        assert_utils.assert_true(resp[0], f"Failed to upload parts. Response: {resp}")
-        LOGGER.info("Listing parts of partial multipart upload")
-        res = s3_mp_test_obj.list_parts(mpu_id, self.bucket_name, self.object_name)
-        assert_utils.assert_true(res[0], res)
-        for part_n in res[1]["Parts"]:
-            assert_utils.assert_list_item(uploading_parts, part_n["PartNumber"])
-        LOGGER.info("Listed parts of partial multipart upload: %s", res[1])
-        LOGGER.info("Step 1: Successfully performed partial multipart upload for %s parts out "
-                    "of total %s", uploading_parts, len(parts))
-        parts = [ele for ele in parts if ele not in uploading_parts]
-
-        LOGGER.info("Step 2: Shutdown %s (K) data pods one by one and start multipart upload for "
-                    "%s size object in %s parts with every data pod shutdown.",
-                    self.kvalue, file_size * const.Sizes.MB, total_parts)
-        for count in range(1, self.kvalue+1):
-            resp = self.ha_obj.delete_kpod_with_shutdown_methods(
-                master_node_obj=self.node_master_list[0], health_obj=self.hlth_master_list[0],
-                down_method=const.RESTORE_DEPLOYMENT_K8S)
-            assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-            pod_name = list(resp[1].keys())[0]
-            pod_data = list()
-            pod_data.append(resp[1][pod_name]['hostname'])  # hostname
-            pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
-            pod_data.append(resp[1][pod_name]['deployment_name'])  # deployment_name
-            self.pod_dict[pod_name] = pod_data
-            self.pod_name_list.append(pod_name)
-            self.restore_pod = self.deploy = True
-            self.restore_method = const.RESTORE_DEPLOYMENT_K8S
-            assert_utils.assert_true(resp[0], "Cluster/Services status is not as expected")
-            LOGGER.info("Deleted %s pod %s by deleting deployment (unsafe)", count, pod_name)
-
-            uploading_parts = self.system_random.sample(parts, 10)
-            LOGGER.info("Step 3: Perform partial multipart upload for %s parts out of total %s",
-                        uploading_parts, len(parts))
-            resp = self.ha_obj.partial_multipart_upload(s3_data=self.s3_clean,
-                                                        bucket_name=self.bucket_name,
-                                                        object_name=self.object_name,
-                                                        part_numbers=uploading_parts,
-                                                        remaining_upload=True, mpu_id=mpu_id,
-                                                        multipart_obj_size=file_size,
-                                                        total_parts=total_parts,
-                                                        multipart_obj_path=object_path)
-
-            assert_utils.assert_true(resp[0], f"Failed to upload parts {resp[1]}")
-            LOGGER.info("Listing parts of partial multipart upload")
-            res = s3_mp_test_obj.list_parts(mpu_id, self.bucket_name, self.object_name)
-            assert_utils.assert_true(res[0], res)
-            for part_n in res[1]["Parts"]:
-                assert_utils.assert_list_item(uploading_parts, part_n["PartNumber"])
-            LOGGER.info("Listed parts of partial multipart upload: %s", res[1])
-            parts_etags.extend(resp[3])
-            parts = [ele for ele in parts if ele not in uploading_parts]
-            LOGGER.info("Step 3: Successfully performed partial multipart upload for %s parts out "
-                        "of total %s", uploading_parts, len(parts))
-        LOGGER.info("Step 2: Successfully shutdown %s (K) data pods one by one and uploaded "
-                    "total %s parts", self.kvalue, uploading_parts)
-
-        LOGGER.info("Step 4: Upload remaining %s parts", parts)
-        resp = self.ha_obj.partial_multipart_upload(s3_data=self.s3_clean,
-                                                    bucket_name=self.bucket_name,
-                                                    object_name=self.object_name,
-                                                    part_numbers=parts,
-                                                    remaining_upload=True, mpu_id=mpu_id,
-                                                    multipart_obj_size=file_size,
-                                                    total_parts=total_parts,
-                                                    multipart_obj_path=object_path)
-
-        assert_utils.assert_true(resp[0], f"Failed to upload remaining parts {resp[1]}")
-        parts_etags.extend(resp[3])
-        LOGGER.info("Listing parts of partial multipart upload")
-        res = s3_mp_test_obj.list_parts(mpu_id, self.bucket_name, self.object_name)
-        assert_utils.assert_true(res[0], res)
-        for part_n in res[1]["Parts"]:
-            assert_utils.assert_list_item(parts, part_n["PartNumber"])
-        LOGGER.info("Listed parts of partial multipart upload: %s", res[1])
-        LOGGER.info("Step 4: Successfully uploaded remaining parts")
-
-        parts_etag = sorted(parts_etags, key=lambda d: d['PartNumber'])
-
-        LOGGER.info("Step 5: Listing parts of multipart upload")
-        res = s3_mp_test_obj.list_parts(mpu_id, self.bucket_name, self.object_name)
-        assert_utils.assert_true(res[0], res)
-        assert_utils.assert_equal(len(res[1]["Parts"]), total_parts)
-        LOGGER.info("Step 5: Listed parts of multipart upload. Count: %s", len(res[1]["Parts"]))
-
-        LOGGER.info("Step 6: Completing multipart upload & verified upload size is %s",
-                    file_size * const.Sizes.MB)
-        res = s3_mp_test_obj.complete_multipart_upload(mpu_id, parts_etag, self.bucket_name,
-                                                       self.object_name)
-        assert_utils.assert_true(res[0], res)
-        res = s3_test_obj.object_list(self.bucket_name)
-        if self.object_name not in res[1]:
-            assert_utils.assert_true(False, res)
-        result = s3_test_obj.object_info(self.bucket_name, self.object_name)
-        obj_size = result[1]["ContentLength"]
-        LOGGER.debug("Uploaded object info for %s is %s", self.bucket_name, result)
-        assert_utils.assert_equal(obj_size, file_size * const.Sizes.MB)
-        LOGGER.info("Step 6: Multipart upload completed & verified upload size is %s",
-                    file_size * const.Sizes.MB)
-
-        LOGGER.info("Step 7: Download the uploaded object and verify checksum")
-        resp = s3_test_obj.object_download(self.bucket_name, self.object_name, download_path)
-        LOGGER.info("Download object response: %s", resp)
-        assert_utils.assert_true(resp[0], resp[1])
-        download_checksum = self.ha_obj.cal_compare_checksum(file_list=[download_path],
-                                                             compare=False)[0]
-        assert_utils.assert_equal(upload_checksum, download_checksum,
-                                  f"Failed to match checksum: {upload_checksum},"
-                                  f" {download_checksum}")
-        LOGGER.info("Matched checksum: %s, %s", upload_checksum, download_checksum)
-        LOGGER.info("Step 7: Successfully downloaded the object and verified the checksum")
-
-        LOGGER.info("ENDED: Test to verify partial multipart upload after each data pod is "
-                    "failed till K data pods and complete upload after all K data pods are failed")
-
-    @pytest.mark.ha
-    @pytest.mark.lc
-    @pytest.mark.tags("TEST-35784")
-    def test_mpu_during_kpods_shutdown(self):
-        """
-        This test tests multipart upload during data pods failure till K pods
-        """
-        LOGGER.info("STARTED: Test to verify multipart upload during data pods failure till K "
-                    "data pods by delete deployment")
-        file_size = HA_CFG["5gb_mpu_data"]["file_size"]
-        total_parts = self.kvalue * 5 + HA_CFG["5gb_mpu_data"]["total_parts"]
-        part_numbers = list(range(1, total_parts + 1))
-        self.system_random.shuffle(part_numbers)
-        output = Queue()
-        parts_etag = list()
-        download_file = self.test_file + "_download"
-        download_path = os.path.join(self.test_dir_path, download_file)
-        event = threading.Event()  # Event to be used to send intimation of pod shutdown
-
-        LOGGER.info("Creating IAM user with name %s", self.s3acc_name)
-        resp = self.rest_obj.create_s3_account(acc_name=self.s3acc_name,
-                                               email_id=self.s3acc_email,
-                                               passwd=S3_CFG["CliConfig"]["s3_account"]["password"])
-        assert_utils.assert_true(resp[0], resp[1])
-        access_key = resp[1]["access_key"]
-        secret_key = resp[1]["secret_key"]
-        s3_test_obj = S3TestLib(access_key=access_key, secret_key=secret_key,
-                                endpoint_url=S3_CFG["s3_url"])
-        s3_mp_test_obj = S3MultipartTestLib(access_key=access_key, secret_key=secret_key,
-                                            endpoint_url=S3_CFG["s3_url"])
-        LOGGER.info("Successfully created IAM user with name %s", self.s3acc_name)
-        self.s3_clean = {'s3_acc': {'accesskey': access_key, 'secretkey': secret_key,
-                                    'user_name': self.s3acc_name}}
-
-        LOGGER.info("Step 1: Start multipart upload of %s object in background",
-                    file_size * const.Sizes.MB)
-        args = {'s3_data': self.s3_clean, 'bucket_name': self.bucket_name,
-                'object_name': self.object_name, 'file_size': file_size, 'total_parts': total_parts,
-                'multipart_obj_path': self.multipart_obj_path, 'part_numbers': part_numbers,
-                'parts_etag': parts_etag, 'output': output}
-        thread = threading.Thread(target=self.ha_obj.start_random_mpu, args=(event,), kwargs=args)
-        thread.daemon = True  # Daemonize thread
-        thread.start()
-        LOGGER.info("Step 1: Started multipart upload of %s object in background",
-                    file_size * const.Sizes.MB)
-        time.sleep(HA_CFG["common_params"]["60sec_delay"])
-
-        LOGGER.info("Step 2: Shutdown random %s (K) data pods by deleting deployment while "
-                    "continuous multipart upload in background and verify cluster & remaining "
-                    "pods status", self.kvalue)
-        resp = self.ha_obj.delete_kpod_with_shutdown_methods(
-            master_node_obj=self.node_master_list[0], health_obj=self.hlth_master_list[0],
-            down_method=const.RESTORE_DEPLOYMENT_K8S, kvalue=self.kvalue)
-        # Assert if empty dictionary
-        assert_utils.assert_true(resp[1], "Failed to shutdown/delete pod")
-        for pod_name in resp[1].keys():
-            pod_data = list()
-            pod_data.append(resp[1][pod_name]['hostname'])  # hostname
-            pod_data.append(resp[1][pod_name]['deployment_backup'])  # deployment_backup
-            pod_data.append(resp[1][pod_name]['deployment_name'])  # deployment_name
-            self.pod_name_list.append(pod_name)
-            self.pod_dict[pod_name] = pod_data
-        self.restore_pod = self.deploy = True
-        self.restore_method = const.RESTORE_DEPLOYMENT_K8S
-        assert_utils.assert_true(resp[0], "Cluster/Services status is not as expected")
-        LOGGER.info("Step 2: Successfully shutdown %s (K) data pod %s. Verified cluster and "
-                    "services states are as expected & remaining pods status is online.",
-                    self.kvalue, self.pod_name_list)
-
-        LOGGER.info("Step 3: Checking response from background process")
-        thread.join()
-        responses = tuple()
-        while len(responses) < 4:
-            responses = output.get(timeout=HA_CFG["common_params"]["60sec_delay"])
-        if not responses:
-            assert_utils.assert_true(False, "Background process failed to do multipart upload")
-        exp_failed_parts = responses[0]
-        failed_parts = responses[1]
-        parts_etag = responses[2]
-        mpu_id = responses[3]
-        LOGGER.debug("Responses received from background process:\nexp_failed_parts: "
-                     "%s\nfailed_parts: %s\nparts_etag: %s\nmpu_id: %s", exp_failed_parts,
-                     failed_parts, parts_etag, mpu_id)
-        if len(exp_failed_parts) == 0 and len(failed_parts) == 0:
-            LOGGER.info("All the parts are uploaded successfully")
-        elif failed_parts:
-            assert_utils.assert_true(False,
-                                     "Failed to upload parts when cluster was in degraded state."
-                                     f"Failed parts: {failed_parts}")
-        elif exp_failed_parts:
-            LOGGER.info("Step 3.1: Upload expected failed remaining parts")
-            resp = self.ha_obj.partial_multipart_upload(s3_data=self.s3_clean,
-                                                        bucket_name=self.bucket_name,
-                                                        object_name=self.object_name,
-                                                        part_numbers=exp_failed_parts,
-                                                        remaining_upload=True,
-                                                        multipart_obj_size=file_size,
-                                                        total_parts=total_parts,
-                                                        multipart_obj_path=self.multipart_obj_path,
-                                                        mpu_id=mpu_id)
-            assert_utils.assert_true(resp[0],
-                                     f"Failed to upload expected failed remaining parts {resp[1]}")
-            parts_etag1 = resp[3]
-            parts_etag = parts_etag + parts_etag1
-            LOGGER.info("Step 3.1: Successfully uploaded expected failed remaining parts")
-        LOGGER.info("Step 3: Successfully checked background process responses")
-
-        parts_etag = sorted(parts_etag, key=lambda d: d['PartNumber'])
-
-        LOGGER.info("Calculating checksum of file %s", self.multipart_obj_path)
-        upload_checksum = self.ha_obj.cal_compare_checksum(file_list=[self.multipart_obj_path],
-                                                           compare=False)[0]
-
-        LOGGER.info("Step 4: Listing parts of multipart upload")
-        res = s3_mp_test_obj.list_parts(mpu_id, self.bucket_name, self.object_name)
-        assert_utils.assert_true(res[0], res)
-        assert_utils.assert_equal(len(res[1]["Parts"]), total_parts)
-        LOGGER.info("Step 4: Listed parts of multipart upload. Count: %s", len(res[1]["Parts"]))
-
-        LOGGER.info("Step 5: Completing multipart upload and check upload size is %s",
-                    file_size * const.Sizes.MB)
-        res = s3_mp_test_obj.complete_multipart_upload(mpu_id, parts_etag, self.bucket_name,
-                                                       self.object_name)
-        assert_utils.assert_true(res[0], res)
-        res = s3_test_obj.object_list(self.bucket_name)
-        assert_utils.assert_in(self.object_name, res[1], res)
-        result = s3_test_obj.object_info(self.bucket_name, self.object_name)
-        obj_size = result[1]["ContentLength"]
-        LOGGER.debug("Uploaded object info for %s is %s", self.bucket_name, result)
-        assert_utils.assert_equal(obj_size, file_size * const.Sizes.MB)
-        LOGGER.info("Step 5: Multipart upload completed and verified upload object size is %s",
-                    obj_size)
-
-        LOGGER.info("Step 6: Download the uploaded object and verify checksum")
-        resp = s3_test_obj.object_download(self.bucket_name, self.object_name, download_path)
-        LOGGER.info("Download object response: %s", resp)
-        assert_utils.assert_true(resp[0], resp[1])
-        download_checksum = self.ha_obj.cal_compare_checksum(file_list=[download_path],
-                                                             compare=False)[0]
-        assert_utils.assert_equal(upload_checksum, download_checksum,
-                                  f"Failed to match checksum: {upload_checksum},"
-                                  f" {download_checksum}")
-        LOGGER.info("Matched checksum: %s, %s", upload_checksum, download_checksum)
-        LOGGER.info("Step 6: Successfully downloaded the object and verified the checksum")
-
-        LOGGER.info("COMPLETED: Test to verify multipart upload during data pods failure till K "
-                    "pods by delete deployment")
